# docs/tms_openapi.yml
openapi: 3.1.0
info:
  title: TMS portal server
  description: TMS portal server documentation
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development
  - url: https://tms-auth-service.tacc.cloud
    description: Development Server
paths:
  "/login":
    post:
      tags:
        - Login
      summary: Login
      description: Login to TMS Portal using specified identity provider, and redirect back to redirect uri (must be |
        configured as an allowed redirect uri)
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TmsLoginRequest'
      responses:
        "307":
          description: Temporary redirect
          headers:
            location:
              schema:
                type: string
              description: redirect url
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TmsErrorResponse'
    get:
      tags:
        - Login
      summary: Login
      description: Login to TMS Portal using specified identity provider, and redirect back to redirect uri (must be |
        configured as an allowed redirect uri)
      parameters:
        - name: idp_id
          schema:
            type: string
          in: query
          required: false
        - name: redirect_uri
          schema:
            type: string
          in: query
          required: false
        - name: client_return_uri
          schema:
            type: string
          in: query
          required: false
        - name: client_name
          schema:
            type: string
          in: query
          required: false
      responses:
        "307":
          description: Temporary redirect
          headers:
            location:
              schema:
                type: string
              description: redirect url
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TmsErrorResponse'
...
  "/login/whoami":
    get:
      tags:
        - Login
      summary: whoami
      security:
        - Bearer: [ ]
      description: Get information about logged in user
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TmsWhoAmIResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TmsErrorResponse'
...
  "/login/callback":
    get:
      tags:
        - Login
      summary: callback
      description: Called by identity provider with code that can be exchanged for a token.
      parameters:
        - name: code
          schema:
            type: string
          in: query
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TmsCallbackResponse'
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TmsErrorResponse:
      type: object
      properties:
        status:
          type: string
        result:
          type: string
    TmsWhoAmIResponse:
      type: object
      properties:
        status:
          type: string
        result:
          properties:
            name:
              type: string
            user_name:
              type: string
            idpDisplayName:
              type: string
            organization:
              type: string
    TmsLoginRequest:
      required:
        - idp_id
        - redirect_uri
        - client_return_uri
        - client_name
      type: object
      properties:
        idp_id:
          type: string
        redirect_uri:
          type: string
        client_return_uri:
          type: string
        client_name:
          type: string
...
    TmsCallbackResponse:
      type: object
      properties:
        status:
          type: string
        result:
          properties:
            token:
              type: string
...